requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.1.7
django-cors-headers==4.7.0
django-phone-field==1.8.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
pillow==11.1.0
PyJWT==2.9.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
sqlparse==0.5.3
typing==3.7.4.3
tzdata==2025.1
uritemplate==4.1.1
